Add  Card::clicked() signal, connect it to Game::onCardClicked
I looked at creating a ClickableGraphicsItem class that could later be used as a base class for
ClickableGraphicsTextItem, but the base class does NOT inherit QObject, while the QGraphicsTextItem
already DOES inherit QObject

Also, having the signal in the base class meant that the QObject::connect() function needed a signal
pointer of the base class type, which resulted in ugly creation of separate base class and derived
class pointers, with associated casts to make everything work.

While adding the clicked event to the Card class is not reusable, it makes the code a lot cleaner.