Add Deck::fanCards to Reposition and Animate cards fanning out across the table
CardStacks become QGraphicsObject to support slots relating to animations
Cards also become QGraphicsObject so we can animate their positions

Note that the animation methods do NOT appear in the base CardStack, this is because the different types
of stacks use different data structures for the cards:  RandomStack is a queue, SortedStack is a list.

This oddity led to a common static method  (getFanLocation) being used instead of a base class method.

Issues:
Some oddities are happening with the stacking after animating the sorting of the deck.
The stack order is essentially random.  Tried several approaches see "algorithm" static int.

Also, QParallelAnimationGroup is not working, not sure why, code looks correct. disabled via a boolean "useGroup"