Fix Undo not restoring card flip state on SortedStacks correctly.
The SortedStacks are more complex when you remove a card - TakeCard will flip over the next card on the stack if it needs to.
This was getting messed up on on undo.

To fix this, the CardStacks now have a "topFlipped" property, which is basically accessed ONLY when calling addCard from an undo method.