Convert UndoStack to a shared pointer, pass to constructor of cardStacks.
The UndoStack was previously used only by the Game Object, but the "dropEvents" are implemented in the CardStack objects.

We now pass a pointer to the UndoStack into the CardStacks when constructing them.

Using the QSharedPtr constructor that takes QObject::deleteLater function as a parameter allows the undo stack to be
destroyed only after all objects are destroyed.