Fix undo move to foundation not restoring top card flip correctly.
I have already developed a method for correctly restoring the state of a top card on an undo, this change
adds that existing method (remember whether you flipped top card when removing the card during the
initial redo, and then use that knowledge on undo.

Also remove some commented code that I started that I never need.
I was able to reuse the existing PlayfieldToFoundation command in the drag
as well as the click handler.