Refactor: Move mCards into base class of CardStacks - significantly reduces duplication of code
Initially I decided NOT to put this in the base class as the Random and Sorted stack can be implemented as true
stacks, while the DescendingStack needs to be able to access below the top/back in order to drag a stack of cards.

However, this lead to a LOT of code duplication. Also, considering that QStack is actually implemented as an overload
of QList, there is no performance advantage to the simpler stack interface.

methods moved to the base class:
addCard
canTake
takeCard (with overloads for positioning)
takeTop (one overload to set flipTop for DescendingStack