Fix card images not being deleted due to Qt ownership not working correctly.
Qt Memory management relies on the parent child relationship to delete child objects
when parents are deleted.  Although the mImage/mBackImage constructors are passed
a parent, the parent was not set, leading to a leak of the card images.

Manually setting the parent after the constructor return fixes this.
See code on parentDebug branch proving this.