Debug code proving that card images are deleted with the card
Previous code (two commits back) should have been correct, as the parent was passed
to the Qt constructor.   However, in reviewing this code, I wanted to double check that the
objects were properly deleted and they were not.      Some debugging indicated that
mImage->parent() was null despite the parent having been sent to the constructor.

Some googling and chatting with AIs suggested the separate setting of parent would
do the trick and it did.